requirements.txt
alembic==1.7.5
apache-airflow==1.10.15
apispec==1.3.3
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
cached-property==1.5.2
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.37
cryptography==36.0.1
defusedxml==0.7.1
dill==0.3.4
dnspython==2.2.0
docutils==0.18.1
email-validator==1.1.3
Flask==1.1.4
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-resources==1.5.0
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.6
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
natsort==8.0.2
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pendulum==1.4.4
prison==0.2.1
psutil==5.9.0
psycopg2-binary==2.9.3
pyarrow==6.0.1
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
requests==2.23.0
setproctitle==1.2.2
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.38.2
tabulate==0.8.9
tenacity==4.12.0
text-unidecode==1.3
thrift==0.15.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
Werkzeug==0.16.1
WTForms==3.0.1
zope.deprecation==4.4.0
